pyproject.toml
[project]
name = "fitcal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-allauth[socialaccount]>=65.7.0",
    "django-environ>=0.12.0",
    "email-validator>=2.2.0",
    "psycopg[binary]>=3.2.7",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.0",
    "ruff>=0.11.8",
    "django-debug-toolbar>=5.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "factory-boy>=3.3.3",
    "pytest-factoryboy>=2.7.0",
]

# ruff設定
[tool.ruff]
# 啟用基本錯誤檢查與 import 排序
select = ["E", "F", "I"]
# 可略過某些資料夾（例如虛擬環境）
exclude = ["venv", ".venv", "__pycache__", "migrations"]
# 行寬與 Python 目標版本
line-length = 88
target-version = "py311"
# 自動修復違規（包含格式與排序）
fix = true
unfixable = ["F401"]

[tool.ruff.format]
# 使用單引號
quote-style = "single"
# 換行符號
line-ending = "lf"

[tool.ruff.isort]
# 排序 import 時合併 as 語法
combine-as-imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "fitcal.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
